move logging utils to a separate .py file
