+++ autointent/context/optimization_info/optimization_info.py
@@ -1,9 +1,7 @@
-import logging
-from pprint import pformat
-
 import numpy as np
 
 from .data_models import Artifact, Artifacts, NDArray, RetrieverArtifact, ScorerArtifact, Trial, Trials, TrialsIds
+from .logger import get_logger
 
 
 class OptimizationInfo:
@@ -80,25 +78,3 @@ class OptimizationInfo:
             "metrics": node_wise_metrics,
             "configs": self.trials.model_dump(),
         }
-
-def get_logger() -> logging.Logger:
-    logger = logging.getLogger(__name__)
-
-    formatter = PPrintFormatter()
-    ch = logging.StreamHandler()
-    ch.setFormatter(formatter)
-    logger.addHandler(ch)
-
-    return logger
-
-
-class PPrintFormatter(logging.Formatter):
-    def __init__(self):
-        super().__init__(fmt="{asctime} - {name} - {levelname} - {message}", style="{")
-
-    def format(self, record):
-        if isinstance(record.msg, dict):
-            format_msg = "module scoring results:\n"
-            dct_to_str = pformat(record.msg)
-            record.msg = format_msg + dct_to_str
-        return super().format(record)
